requirements.txt
grpcio
grpcio-tools
protobuf
redis
fastapi
uvicorn
uvloop